+++ lib/uix.dart
@@ -11,6 +11,7 @@ export 'src/scheduler.dart';
 export 'src/env.dart';
 export 'src/data/node.dart';
 export 'src/vdom/attr.dart';
+export 'src/vdom/prop.dart';
 export 'src/vdom/style.dart';
 export 'src/vdom/vcontext.dart';
 export 'src/vdom/anchor.dart';